(Not tested) Added UserAuthButton
- Wraps Login and Signup buttons
