+++ frontend/src/Components/navigationbar/settingsbutton/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SettingsButton = () => {
+  return <div></div>;
+};
+
+export default SettingsButton;
