--- frontend/src/Components/navigationbar/signupbutton/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SignupButton = () => {
-  return <div></div>;
-};
-
-export default SignupButton;
