--- frontend/src/Components/navigationbar/userauth/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const UserAuth = () => {
-  return <div></div>;
-};
-
-export default UserAuth;
